OOP section all tests passing except for nullary constructor comparisons
something off with all of the comparisons that try to compare two empty addresses.  The expected value ends up with an extra space outside of the <> that define the toString result.  I ran into this once before, and Wilhelm figured out in that case it had to do with some field being set that wasn't included in the toString.  I will try again to step through the objects in debug mode to see if I can figure out the difference.